% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdakma-package.R
\docType{package}
\name{fdakma-package}
\alias{fdakma}
\alias{fdakma-package}
\title{fdakma: Functional Data Analysis: K-Mean Alignment}
\description{
It performs simultaneously clustering and alignment of a multidimensional or unidimensional functional dataset by means of k-mean alignment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/astamm/fdakma}
  \item Report bugs at \url{https://github.com/astamm/fdakma/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID}) [contributor]

Authors:
\itemize{
  \item Laura Sangalli \email{laura.sangalli@polimi.it}
  \item Piercesare Secchi \email{piercesare.secchi@polimi.it}
  \item Simone Vantini \email{simone.vantini@polimi.it}
  \item Valeria Vitelli \email{valeria.vitelli@medisin.uio.no}
}

Other contributors:
\itemize{
  \item Alessandro Zito \email{zito.ales@gmail.com} [contributor]
}

}
\keyword{internal}
